refactor: MyReservations(538行) を責務ごとに分割（Phase 2）
- hooks/use-my-reservation-groups.ts: グループ化・カレンダー派生・セクション分け・延滞集計・badgeOf を集約
- _components/ReservationCard.tsx: 1グループ分の予約カード（機材行＋一括操作）
- _components/CancelReturnDialogs.tsx: キャンセル・返却の確認ダイアログ
本体は約200行に縮小。振る舞いは不変（build/test green）。
